+++ packages/keepkit-ui/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.tsx"],
+  entry: ["src/index.tsx", "src/tailwind.ts"],
   format: ["esm"],
   dts: true,
   sourcemap: true,